+++ packages/solid/src/SolidWIndowSplitter.tsx
